# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zennex API Documentation",
        "version": "1.0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь успешно авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильные учетные данные"
                    }
                }
            }
        },
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Получение списка заметок",
                "description": "Возвращает список всех заметок",
                "operationId": "9a95bde5931986ad60b9f7f414ec425e",
                "responses": {
                    "200": {
                        "description": "Список заметок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Note"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заметки не найдены"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Создать новую заметку",
                "operationId": "d133c170557b4935898f2e444e0fd418",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Новая заметка"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Содержание заметки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заметка создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Получить заметку по ID",
                "operationId": "4f26c25d4040d0ef0fa8d8a5ea9fa4da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заметка найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Удалить заметку",
                "operationId": "99b63679b2822651f8f0ae585f927f0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заметки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Заметка успешно удалена"
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                }
            }
        },
        "/notes/{id}": {
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Обновление заметки",
                "description": "Обновляет существующую заметку и синхронизирует теги.",
                "operationId": "29a9163f38861fad93ae0510c516c5cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заметки для обновления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Заголовок заметки",
                                        "type": "string"
                                    },
                                    "content": {
                                        "description": "Содержимое заметки",
                                        "type": "string"
                                    },
                                    "tags": {
                                        "description": "Массив тегов",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление заметки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Заметка не найдена"
                    }
                }
            }
        },
        "/notes/search-by-tags": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Поиск заметок по тегам",
                "description": "Возвращает список заметок, соответствующих переданным тегам.",
                "operationId": "0e11a7fe1fb54fff2cbbaa6c93836119",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Массив тегов для поиска заметок",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заметок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Note"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос"
                    }
                }
            }
        },
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Получение списка тегов",
                "description": "Возвращает список всех тегов",
                "operationId": "4d0f4baf9e6fa992c0d85fbf140ebf17",
                "responses": {
                    "200": {
                        "description": "Список тегов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Теги не найдены"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Create a new tag",
                "operationId": "a0a246f29dcb119a2189fd230fc10028",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tag Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag created successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Note": {
                "required": [
                    "title",
                    "content",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID заметки",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Название заметки",
                        "type": "string"
                    },
                    "content": {
                        "description": "Контент заметки",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID пользователя создавшего заметку",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Метка даты",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Последнее обновление",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "title": "Tag",
                "description": "Модель тега",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Идентификатор тега",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Название тега",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID пользователя, которому принадлежит тег",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Notes",
            "description": "Notes"
        },
        {
            "name": "Tags",
            "description": "Tags"
        }
    ]
}